+++ arenatacticsserver/scenes/actions/offensive_actions/basic_offensive_action.gd
@@ -5,7 +5,7 @@ class_name BasicOffensiveAction
 @export var actor_effect: OffensiveActionEffect = null
 @export var default_offensive_icon: Texture2D = preload("res://scenes/actions/offensive_actions/assets/offensive_icon_placeholder.png")
 
-func _init(instance: CharacterInstance, _data: Dictionary, _name: String):
+func _init(instance: CharacterInstance = null, _data: Dictionary = {}, _name: String = "Default"):
 	self.icon = default_offensive_icon
 	super(instance, _data, _name)
 	for key in _data: