+++ arenatacticsserver/scenes/actions/offensive_actions/melee_offensive_action.gd
@@ -1,5 +1,5 @@
 extends BasicOffensiveAction
 class_name MeleeOffensiveAction
 
-func _init(instance: CharacterInstance, _data: Dictionary, _name: String):
+func _init(instance: CharacterInstance = null, _data: Dictionary = {}, _name: String = "Default"):
 	super(instance, _data, _name)